Adjust the code paths under the examples/ directory to prevent the redefinition of main.
